requirements.txt
streamlit==1.32.0
streamlit-autorefresh==1.0.1
anthropic>=0.28.0
supabase>=2.4.0
plotly==5.20.0
pandas==2.2.1
python-telegram-bot==21.0
pytz==2024.1
python-dotenv==1.0.1
extra-streamlit-components==0.1.71
